requirements.txt
python3.6
pytorch
gym
numpy
PIL
argparse
tensorboardX
